# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Calculator Service
  description: Service for adding numbers, a Goa teaser
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /add/{a}/{b}:
    get:
      tags:
      - calc
      summary: add calc
      operationId: calc#add
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /division/{a}/{b}:
    get:
      tags:
      - calc
      summary: division calc
      operationId: calc#division
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /multiplicacion/{a}/{b}:
    get:
      tags:
      - calc
      summary: multiplicacion calc
      operationId: calc#multiplicacion
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /openapi.json:
    get:
      tags:
      - calc
      summary: Download ./gen/http/openapi.json
      operationId: calc#/openapi.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
  /resta/{a}/{b}:
    get:
      tags:
      - calc
      summary: resta calc
      operationId: calc#resta
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        type: integer
      - name: b
        in: path
        description: Right operand
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http